Several bugfixes - Added cooldown, stamina, and mana cost buffs to stats - Fixed mortal engine, practice dummy null pointer - Added a way to check for nullification in receiveddamageevent - Fixed glossary inventories disabling player session inventory on return to inv - Fixed shrine instance board - Fixed damage meta not copying fully - Fixed sanctified stat

+++ src/me/neoblade298/neorogue/session/ShrineInstance.java
@@ -99,7 +99,7 @@ public class ShrineInstance extends EditInventoryInstance {
 		if (isHost) {
 			line += "(Host) ";
 		}
-		line += data.getData().getDisplay() + " (" + Math.round(data.getHealth()) + " / " + data.getMaxHealth()
+		line += data.getData().getDisplay() + " (" + Math.round(data.getHealth()) + " / " + (int) data.getMaxHealth()
 				+ "§c♥§f)";
 		return line;
 	}
